requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
chromadb==0.4.22
openai==1.10.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
tiktoken==0.5.2
aiofiles==23.2.1
sqlalchemy==2.0.25
alembic==1.13.1
rich==13.7.0
pytest==7.4.3
httpx==0.26.0
